serverstack: introduce the package; relocate backend dispatch
New internal/serverstack package — the future single construction path
for graph/indexer/engine/MCP, kept out of internal/daemon to avoid the
daemon->mcp->daemon import cycle. The backend dispatch (memory|sqlite +
path resolution + the pure-Go sqlite open) moves here as OpenBackend;
cmd/gortex keeps a thin openBackend wrapper so existing call sites
compile unchanged. Backend-dispatch tests added.
